requirements.txt
asgiref==3.2.7
Django==3.0.6
django-restframework==0.0.1
djangorestframework==3.11.0
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.3.1
sqlparse==0.3.1
